+++ src/psimage.c
@@ -461,7 +461,7 @@ int GMT_psimage (void *V_API, int mode, void *args) {
 
 			/* Expand 8-bit indexed image to 24-bit image */
 			I->data = gmt_M_memory (GMT, I->data, 3 * I->header->nm, unsigned char);
-			n = (int)(3 * I->header->nm - 1);
+			n = 3 * I->header->nm - 1;
 			for (j = (int)I->header->nm - 1; j >= 0; j--) {
 				k = 4 * I->data[j] + 3;
 				if (has_trans && colormap[k] == 0)
@@ -473,7 +473,7 @@ int GMT_psimage (void *V_API, int mode, void *args) {
 		}
 		else if (I->header->n_bands == 4) { /* RGBA image, with a color map */
 			uint64_t n4, j4;
-			if (!Ctrl->G.active) has_trans = find_unique_color (GMT, I->data, (int)I->header->nm, &r, &g, &b);
+			if (!Ctrl->G.active) has_trans = find_unique_color (GMT, I->data, I->header->nm, &r, &g, &b);
 			for (j4 = n4 = 0; j4 < 4 * I->header->nm; j4++) { /* Reduce image from 32- to 24-bit */
 				if (has_trans && I->data[j4+3] == 0)
 					I->data[n4++] = (unsigned char)r, I->data[n4++] = (unsigned char)g, I->data[n4++] = (unsigned char)b, j4 += 3;
